modiff

+++ src/main/java/tn/esprit/pidev/repositories/CondidacyRepository.java
@@ -4,8 +4,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import tn.esprit.pidev.entities.Condidacy;
 
+import java.util.List;
+
 @Repository
-public interface CondidacyRepository extends JpaRepository<Condidacy, Long> {
+public interface CondidacyRepository extends JpaRepository<Condidacy,Integer> {
+
